+++ Fusion/Algorithms/Ecofer/Data/MINP.cs
@@ -43,9 +43,9 @@ namespace Data
             MINP_ProcessPattern = null;
         }
 
-        public static float Sum(float[] aArray, int aFromIndex, int aToIndex)
+        public static double Sum(double[] aArray, int aFromIndex, int aToIndex)
         {
-            float lResult = 0;
+            double lResult = 0;
 
             for (int i = aFromIndex; i <= aToIndex; i++)
             {
@@ -54,39 +54,39 @@ namespace Data
 
             return lResult;
         }
-        public static float FP(DTO.MINP_GD_MaterialDTO aMaterial, int aIndex)
+        public static double FP(DTO.MINP_GD_MaterialDTO aMaterial, int aIndex)
         {
             if (aMaterial == null) return 0;
             DTO.MINP_GD_MaterialItemsDTO lMaterialItem = aMaterial.MINP_GD_MaterialItems.SingleOrDefault(aR => aR.MINP_GD_MaterialElement.Index == aIndex);
-            return lMaterialItem == null ? 0 : (float)lMaterialItem.Amount_p;
+            return lMaterialItem == null ? 0 : (double)lMaterialItem.Amount_p;
         }
-        public static float ConversionVector(int aIndex)
+        public static double ConversionVector(int aIndex)
         {
-            return (float)MINP.MINP_GD_MaterialElements[aIndex].Vector;
+            return (double)MINP.MINP_GD_MaterialElements[aIndex].Vector;
         }
-        public static float Mm(int aIndex)
+        public static double Mm(int aIndex)
         {
-            return (float)MINP.MINP_GD_MaterialElements[aIndex].Mm;
+            return (double)MINP.MINP_GD_MaterialElements[aIndex].Mm;
         }
-        public static float O2_Stechio(int aIndex)
+        public static double O2_Stechio(int aIndex)
         {
-            return (float)MINP.MINP_GD_MaterialElements[aIndex].O2;
+            return (double)MINP.MINP_GD_MaterialElements[aIndex].O2;
         }
-        public static float E_Ox1(int aIndex)
+        public static double E_Ox1(int aIndex)
         {
-            return (float)MINP.MINP_GD_MaterialElements[aIndex].E_Ox1;
+            return (double)MINP.MINP_GD_MaterialElements[aIndex].E_Ox1;
         }
-        public static float E_Ox2(int aIndex)
+        public static double E_Ox2(int aIndex)
         {
-            return (float)MINP.MINP_GD_MaterialElements[aIndex].E_Ox2;
+            return (double)MINP.MINP_GD_MaterialElements[aIndex].E_Ox2;
         }
-        public static float Eta_Ox1(int aIndex)
+        public static double Eta_Ox1(int aIndex)
         {
-            return (float)MINP.MINP_GD_MaterialElements[aIndex].Eta_Ox1;
+            return (double)MINP.MINP_GD_MaterialElements[aIndex].Eta_Ox1;
         }
-        public static float Eta_Ox2(int aIndex)
+        public static double Eta_Ox2(int aIndex)
         {
-            return (float)MINP.MINP_GD_MaterialElements[aIndex].Eta_Ox2;
+            return (double)MINP.MINP_GD_MaterialElements[aIndex].Eta_Ox2;
         }
     }
 }
